backend/requirements.txt
fastapi
uvicorn
yfinance
python-dotenv
langchain
google-generativeai
pandas
numpy
